BinTree.Internal

Description

Module name: Bintree.Internal Description: our own Haskell implementation of a Binary Search Trees, provides basic (common) features of BinTree and AVLTree.

Documentation

data BinTree a

Definition of BinTree data type.

Constructors

EmptyBT
Empty node

BTNode a (BinTree a) (BinTree a)
Node with a left and a right son

height :: (Num b, Ord b) => BinTree a -> b

Returns height of the tree.

left :: BinTree a -> BinTree a

Returns value of node's left son.

right :: BinTree a -> BinTree a

Returns value of node's right son.

value :: Num a => BinTree a -> a

Returns node's value (assuming that empty BinTree has no value!)

getValue :: Maybe a -> a

Gets node's value (attempt to get value of empty BinTree calls an error).

search :: (Num a, Ord a) => a -> BinTree a -> Bool

Checks if a given value exists in a BinTree.

inorder :: BinTree a -> [a]

Performs an inorder conversion to simple list, for example: inorder BTNode 4 (BTNode 2 (BTNode 1 EmptyBT EmptyBT) (BTNode 3 EmptyBT EmptyBT)) (BTNode 5 EmptyBT (BTNode 6 EmptyBT (BTNode 7 EmptyBT EmptyBT))) => [1, 2, 3, 4, 5, 6, 7]

postorder :: BinTree a -> [a]

Performs a postorder conversion to simple list, for example: postorder BTNode 4 (BTNode 2 (BTNode 1 EmptyBT EmptyBT) (BTNode 3 EmptyBT EmptyBT)) (BTNode 5 EmptyBT (BTNode 6 EmptyBT (BTNode 7 EmptyBT EmptyBT))) => [1, 3, 2, 7, 6, 5, 4]

preorder :: BinTree a -> [a]

Performs a preorder conversion to simple list, for example: preorder BTNode 4 (BTNode 2 (BTNode 1 EmptyBT EmptyBT) (BTNode 3 EmptyBT EmptyBT)) (BTNode 5 EmptyBT (BTNode 6 EmptyBT (BTNode 7 EmptyBT EmptyBT))) => [4, 2, 1, 3, 5, 6, 7]

ifEqBinTrees :: Eq a => BinTree a -> BinTree a -> Bool

Checks if two BinTrees are equal, which means that they contain the same values.