Dequeue

Documentation

data Dequeue a

Dequeue data structure.

emptyDEQ :: Dequeue a

Creates empty Dequeue.

isEmptyDEQ :: Dequeue a -> Bool

Checks whether Dequeue is empty.

lengthDEQ :: Dequeue a -> Int

Returns length of a Dequeue.

firstDEQ :: Dequeue a -> Maybe a

Returns first element of a Dequeue (Maybe a, where a is a type of data stored in Dequeue) or Nothing, if it doesn't exist.

lastDEQ :: Dequeue a -> Maybe a

Returns last element of a Dequeue (Maybe a, where a is a type of data stored in Dequeue) or Nothing, if it doesn't exist.

takeFrontDEQ :: Int -> Dequeue a -> [a]

takeFrontDEQ n, applied to a Dequeue d, returns the prefix of d of length n, or list consisiting of all values of d in original order if n > lengthDEQ d.

takeBackDEQ :: Int -> Dequeue a -> [a]

takeBackDEQ n, applied to a Dequeue d, returns the last n values of d as a list, or all of d's values reversed as a list if n > lengthDEQ d.

pushFrontDEQ :: Dequeue a -> a -> Dequeue a

pushFrontDEQ q a – appends provided value a in the beginning of a returned Dequeue.

popFrontDEQ :: Dequeue a -> Maybe (a, Dequeue a)

popFrontDEQ q – returns tuple consisting of a first element of a Dequeue q and the Dequeue without this element

pushBackDEQ :: Dequeue a -> a -> Dequeue a

pushBackDEQ q a – appends provided value a in the end of a returned Dequeue.

popBackDEQ :: Dequeue a -> Maybe (a, Dequeue a)

popBackDEQ q – returns tuple consisting of a last element of a Dequeue q and the Dequeue without this element

fromListDEQ :: [a] -> Dequeue a

fromListDEQ l – Creates Dequeue from list l.