MyList

Description

Module name: MyList Description: our own implementation of a list in Haskell

Documentation

data MyList a

Definition of our data type.

listHead :: MyList a -> a

Returns head of our list.

listTail :: MyList a -> MyList a

Returns tail of the list (all the elements except for its head).

listFind :: Eq a => a -> MyList a -> Bool

Checks if a given value exists in the list.

myListFromList :: [a] -> MyList a

Builds and returns list of type MyList on the basis of a regular Haskell Data.List.

myListToList :: MyList a -> [a]

Builds a Data.List list on the basis of MyList list

listLength :: MyList a -> Int

Returns length of a given list.

lastElem :: MyList a -> a

Returns last elements of a given list.

elemAtInd :: Int -> MyList a -> a

Looks for an element of a given index in the list and returns it if it exists.

listInsert :: a -> MyList a -> MyList a

Inserts a single element into a given list.

insertInOrder :: Ord a => a -> MyList a -> MyList a

Inserts a single element into a given list, assuming that the given list is sorted.

endInsert :: a -> MyList a -> MyList a

Inserts a single element onto the end of a given list (except for an EmptyList, where it creates a new node at the beginning).

listReverse :: MyList a -> MyList a

Reverses whole MyList list. | Example (simplification): listReverse [1, 2, 3, 4, 5] => [5, 4, 3, 2, 1]

listConcat :: MyList a -> MyList a -> MyList a

Connects two MyList lists together.

mapList :: (a -> b) -> MyList a -> MyList b

Performs MyList's mapping, for example (simplification): | mapList (^2) [1, 2, 3, 4, 5] => [1, 4, 9, 16, 25]

sortList :: Ord a => MyList a -> MyList a

Performs insertion sort on a MyList.