Stack

Documentation

data Stack a

Stack data structure.

empty :: Stack a

Creates empty Stack.

isEmpty :: Stack a -> Bool

Checks whether Stack is empty.

push :: a -> Stack a -> Stack a

push x s – put a value x on a top of a Stack s.

top :: Stack a -> a

top s – returns a value from a top of a Stack s.

pop :: Stack a -> (a, Stack a)

pop s – returns a tuple consisting of a value from a top of a Stack s and the remaining part of a Stack s.